pyproject.toml
[project]
name = "social-alchemist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.52.0",
    "aiohttp",
    "pyaudio",
    "opencv-python",
    "even-glasses",
    "python-dotenv",
    "webrtcvad-wheels>=2.0.14",
    "openwakeword>=0.6.0",
    "onnxruntime>=1.24.2",
]
